+++ validator/tests/test_validation.py
@@ -13,6 +13,8 @@ from dateutil.tz import tzlocal
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 
+from validator.validation.validation import _compare_validation_runs
+
 User = get_user_model()
 from django.test import TestCase
 from django.test.utils import override_settings
@@ -37,7 +39,7 @@ import validator.validation as val
 from validator.validation.batches import _geographic_subsetting
 from validator.validation.globals import METRICS, TC_METRICS
 from validator.validation.globals import OUTPUT_FOLDER
-from api.views.validation_config_view import _compare_validation_runs
+from api.views.validation_config_view import compare_validation_runs
 from api.views.validation_run_view import _copy_validationrun
 from django.shortcuts import get_object_or_404
 from validator.tests.auxiliary_functions import generate_default_validation, generate_default_validation_triple_coll
@@ -1325,7 +1327,7 @@ class TestValidation(TestCase):
                 run.anomalies_to = time_intervals_to
             run.scaling_method = scaling_methods[i]
             run.save()
-            is_there_one = _compare_validation_runs(run, published_runs, user)
+            is_there_one = compare_validation_runs(run, published_runs, user)
 
             assert is_there_one['is_there_validation']
             assert is_there_one['val_id'] == run_ids[i]
@@ -1347,7 +1349,7 @@ class TestValidation(TestCase):
         run.scaling_method = scaling_methods[0]
         run.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # getting back to good coordinates
@@ -1357,26 +1359,26 @@ class TestValidation(TestCase):
         run.max_lon = self.hawaii_coordinates[3]
 
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
         assert is_there_one['val_id'] == run_ids[0]
 
         # spoiling time span:
         run.interval_from = datetime(1990, 1, 1, tzinfo=UTC)
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         run.interval_from = time_intervals_from
         run.interval_to = datetime(2000, 1, 1, tzinfo=UTC)
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # time span restored
         run.interval_to = time_intervals_to
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
         assert is_there_one['val_id'] == run_ids[0]
 
@@ -1384,14 +1386,14 @@ class TestValidation(TestCase):
         # scaling method at the same time):
         run.anomalies = anomalies_methods[1][0]
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         run.anomalies = anomalies_methods[0][0]
         # there is no run with scaling method LINREG
         run.scaling_method = ValidationRun.LINREG
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # restoring existing validation
@@ -1400,26 +1402,26 @@ class TestValidation(TestCase):
         run.anomalies_from = time_intervals_from
         run.anomalies_to = time_intervals_to
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
         assert is_there_one['val_id'] == run_ids[2]
 
         # messing up with anomalies time interval:
         run.anomalies_from = datetime(1990, 1, 1, tzinfo=UTC)
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         run.anomalies_from = time_intervals_from
         run.anomalies_to = datetime(1990, 1, 1, tzinfo=UTC)
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # getting back to appropriate settings
         run.anomalies_to = time_intervals_to
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # getting back to settings of the run with filters set adding filters for the run
@@ -1428,7 +1430,7 @@ class TestValidation(TestCase):
         run.anomalies_from = None
         run.anomalies_to = None
         run.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # adding filters
@@ -1448,7 +1450,7 @@ class TestValidation(TestCase):
         new_pfilter = ParametrisedFilter(filter=DataFilter.objects.get(name="FIL_ISMN_DEPTH"), parameters="0.0,0.1", \
                                          dataset_config=run.reference_configuration)
         new_pfilter.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
 
         assert is_there_one['is_there_validation']
         assert is_there_one['val_id'] == run_filt_id
@@ -1461,7 +1463,7 @@ class TestValidation(TestCase):
             if new_config.dataset.short_name == globals.C3S:
                 new_config.filters.add(c3s_filter)
                 new_config.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
 
         assert not is_there_one['is_there_validation']
 
@@ -1469,7 +1471,7 @@ class TestValidation(TestCase):
         for new_config in run.dataset_configurations.all():
             if new_config.dataset.short_name == globals.C3S:
                 new_config.filters.remove(c3s_filter)
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # removing ismn filter:
@@ -1478,7 +1480,7 @@ class TestValidation(TestCase):
             if new_config.dataset.short_name == globals.ISMN:
                 new_config.filters.remove(ismn_filter)
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # restoring the filter:
@@ -1486,7 +1488,7 @@ class TestValidation(TestCase):
             if new_config.dataset.short_name == globals.ISMN:
                 new_config.filters.add(ismn_filter)
             new_config.save()
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # messing up with parameterised filters:
@@ -1496,7 +1498,7 @@ class TestValidation(TestCase):
                 pf.parameters = 'SCAN,OZNET'
                 pf.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         for pf in ParametrisedFilter.objects.filter(dataset_config=run.reference_configuration):
@@ -1504,7 +1506,7 @@ class TestValidation(TestCase):
                 pf.parameters = 'OZNET'
                 pf.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # restoring networks
@@ -1513,7 +1515,7 @@ class TestValidation(TestCase):
                 pf.parameters = 'SCAN'
                 pf.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # with depths
@@ -1522,7 +1524,7 @@ class TestValidation(TestCase):
                 pf.parameters = '0.10,0.20'
                 pf.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # restoring depths
@@ -1531,7 +1533,7 @@ class TestValidation(TestCase):
                 pf.parameters = '0.0,0.1'
                 pf.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # adding a new dataset:
@@ -1542,12 +1544,12 @@ class TestValidation(TestCase):
         data_c.variable = DataVariable.objects.get(short_name='ASCAT_sm')
         data_c.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         data_c.delete()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert is_there_one['is_there_validation']
 
         # checking scaling reference:
@@ -1555,7 +1557,7 @@ class TestValidation(TestCase):
         run.scaling_ref = new_ref
         run.save()
 
-        is_there_one = _compare_validation_runs(run, published_runs, user)
+        is_there_one = compare_validation_runs(run, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         # ================== tcols ====================================
@@ -1571,7 +1573,7 @@ class TestValidation(TestCase):
         run_tcol.anomalies = anomalies_methods[0][0]
         run_tcol.scaling_method = scaling_methods[0]
         run_tcol.save()
-        is_there_one = _compare_validation_runs(run_tcol, published_runs, user)
+        is_there_one = compare_validation_runs(run_tcol, published_runs, user)
 
         assert is_there_one['is_there_validation']
         assert is_there_one['val_id'] == run_tcol_id
@@ -1580,7 +1582,7 @@ class TestValidation(TestCase):
         run_tcol.tcol = False
         run_tcol.save()
 
-        is_there_one = _compare_validation_runs(run_tcol, published_runs, user)
+        is_there_one = compare_validation_runs(run_tcol, published_runs, user)
         assert not is_there_one['is_there_validation']
 
         ValidationRun.objects.all().delete()
@@ -1619,7 +1621,7 @@ class TestValidation(TestCase):
 
         validations = ValidationRun.objects.exclude(pk=copied_run_info['run_id'])
         copied_run = ValidationRun.objects.get(pk=copied_run_info['run_id'])
-        comparison = _compare_validation_runs(copied_run, validations, copied_run.user)
+        comparison = compare_validation_runs(copied_run, validations, copied_run.user)
 
         # the query validations will be empty so 'is_there_validation' == False, 'val_id' == None, '
         # 'belongs_to_user'==False, 'is_published' == False
@@ -1633,7 +1635,7 @@ class TestValidation(TestCase):
 
         validations = ValidationRun.objects.exclude(pk=copied_run_info['run_id'])
         copied_run = ValidationRun.objects.get(pk=copied_run_info['run_id'])
-        comparison = _compare_validation_runs(copied_run, validations, copied_run.user)
+        comparison = compare_validation_runs(copied_run, validations, copied_run.user)
 
         assert comparison['is_there_validation']
         assert comparison['val_id'] == run.id
